Added "build" folder to project
Added a "build" folder to the root directory of the project. It contains an optimized, minified, compiled, production-ready build of the React JavaScript source code I have written so far, including the Components, classes, and CSS stylesheets from the "src" folder. The web pages of the networking and account system are now viewable from the "index.html" file in the "build" folder without requiring that a new React project be created and the "src" folder imported into it.