Added "build" folder instructions to README.md
Updated the description in the repository's README.md file to include instructions for using the new "build" folder inside the root directory of the repository. Also updated the outdated description stating that a production-ready version of the code was not included in the repository, since the "build" folder now contains this. Clarified the differences between the two versions of the project's source code in the "src" folder and the "build" folder.